+++ database/migrations/2018_08_07_132515_create_tbl_category.php
@@ -14,11 +14,11 @@ class CreateTblCategory extends Migration
     public function up()
     {
         Schema::create('categories', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name')->nullable();
             $table->string('categoriable_type');
             $table->integer('categoriable_id')->unsigned();
-            $table->unsignedInteger('parent_id');
+            $table->unsignedBigInteger('parent_id');
             $table->timestamps();
         });
     }
