+++ database/migrations/2019_10_05_212848_create_departments_table.php
@@ -14,10 +14,10 @@ class CreateDepartmentsTable extends Migration
     public function up()
     {
         Schema::create('departments', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('description')->nullable();
-            $table->unsignedInteger('restaurant_id');
+            $table->unsignedBigInteger('restaurant_id');
             $table->timestamps();
         });
     }
